refactor: remove Tauri dependency from database module
- Move DatabaseDelegate impl for AppHandle from database.rs to lib.rs
- Remove tauri imports from database module
- Achieve complete separation between database logic and Tauri framework
