+++ src-tauri/src/database.rs
@@ -6,31 +6,12 @@ use chrono::Utc;
 use sqlx::Sqlite;
 use sqlx::migrate::MigrateDatabase;
 use sqlx::sqlite::SqlitePool;
-use tauri::{AppHandle, Emitter};
 
 pub trait DatabaseDelegate: Send + Sync {
     fn on_task_created(&self, task: &Task) -> Result<(), Error>;
     fn on_task_popped(&self, task: &Task) -> Result<(), Error>;
     fn on_task_updated(&self, task: &Task) -> Result<(), Error>;
     fn on_stack_updated(&self, stack: &TaskStack) -> Result<(), Error>;
-}
-
-impl DatabaseDelegate for AppHandle {
-    fn on_task_created(&self, task: &Task) -> Result<(), Error> {
-        Ok(self.emit("task:created", task)?)
-    }
-
-    fn on_task_popped(&self, task: &Task) -> Result<(), Error> {
-        Ok(self.emit("task:popped", task)?)
-    }
-
-    fn on_task_updated(&self, task: &Task) -> Result<(), Error> {
-        Ok(self.emit("task:updated", task)?)
-    }
-
-    fn on_stack_updated(&self, stack: &TaskStack) -> Result<(), Error> {
-        Ok(self.emit("stack:updated", stack)?)
-    }
 }
 
 pub struct Database {
